+++ tests/Gedmo/Tree/MaterializedPathORMRepositoryTest.php
@@ -23,6 +23,16 @@ class MaterializedPathORMRepositoryTest extends BaseTestCaseORM
     /** @var \Gedmo\Tree\Entity\Repository\MaterializedPathRepository */
     protected $repo;
 
+    /**
+     * @var TreeListener
+     */
+    private $listener;
+
+    /**
+     * @var array
+     */
+    private $config = [];
+
     protected function setUp(): void
     {
         parent::setUp();